Moved notifications into their own component.
