+++ resources/assets/sass/app.scss
@@ -15,9 +15,3 @@
 
 // Snotify
 @import "~vue-snotify/styles/material";
-.snotifyToast__body{
-  font-size: 1.1em;
-}
-.snotify-success .snotifyToast__body{
-  color: #e1f1e1;
-}
